requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
alembic
psycopg2-binary
slowapi
python-dotenv
pyjwt
passlib
bcrypt==4.0.1
python-multipart
email-validator
Authlib
pydantic
pydantic-settings